+++ app/api/like_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify
 from flask_login import current_user, login_required
-from app.models import db, Like
+from app.models import db, Like, List
 from app.api.auth_routes import validation_errors_to_error_messages
 
 like_routes = Blueprint('likes', __name__)
@@ -12,6 +12,9 @@ def likes(list_id):
     """
     Query for all likes by list id and returns them in a list of like dictionaries
     """
+    list = List.query.get(list_id)
+    if not list:
+        return {'errors': f"List {list_id} does not exist."}, 400
     likes = Like.query.filter(Like.list_id == list_id).all()
     return {'likes': [like.to_dict() for like in likes]}
 
@@ -22,6 +25,12 @@ def like_list(list_id):
     """
     Creates a like
     """
+    list = List.query.get(list_id)
+    if not list:
+        return {'errors': f"List {list_id} does not exist."}, 400
+    existing_like = Like.query.filter((Like.user_id == current_user.id) & (Like.list_id == list_id)).first()
+    if existing_like:
+        return {'message': f"User already likes list {list_id}."}
     like = Like(
         user_id=current_user.id,
         list_id=list_id
@@ -37,6 +46,12 @@ def unlike_list(list_id):
     """
     Deletes a like
     """
+    list = List.query.get(list_id)
+    if not list:
+        return {'errors': f"List {list_id} does not exist."}, 400
+    existing_like = Like.query.filter((Like.user_id == current_user.id) & (Like.list_id == list_id)).first()
+    if not existing_like:
+        return {'message': f"User already does not like list {list_id}."}
     like = Like.query.filter((Like.user_id == current_user.id) & (Like.list_id == list_id)).first()
     db.session.delete(like)
     db.session.commit()
